tezos_interop: integrates fetch_storage (#500)
* tezos_interop: merge fetch_storage in js bridge

* tezos_interop: improve error messages

* tezos_interop: fetch_storage through js bridge
