+++ src/tezos_interop/scripts.ml
@@ -1,6 +1,5 @@
 open Helpers
 
-let fetch_storage = [%blob "fetch_storage.bundle.js"]
 let tezos_js_bridge = [%blob "tezos_js_bridge.bundle.js"]
 let listen_transactions = [%blob "listen_transactions.bundle.js"]
 
@@ -9,13 +8,10 @@ let make_file ~prefix content =
   let%await () = Lwt_io.write oc content in
   await file
 
-let file_fetch_storage, file_tezos_js_bridge, file_listen_transactions =
+let file_tezos_js_bridge, file_listen_transactions =
   Lwt_main.run
-    (let%await file_fetch_storage =
-       make_file ~prefix:"fetch_storage" fetch_storage in
-     let%await file_tezos_js_bridge =
+    (let%await file_tezos_js_bridge =
        make_file ~prefix:"tezos_js_bridge" tezos_js_bridge in
      let%await file_listen_transactions =
        make_file ~prefix:"listen_transactions" listen_transactions in
-     Lwt.return
-       (file_fetch_storage, file_tezos_js_bridge, file_listen_transactions))
+     Lwt.return (file_tezos_js_bridge, file_listen_transactions))
