+++ src/tezos_interop/scripts.mli
@@ -1,3 +1,2 @@
-val file_fetch_storage : string
 val file_tezos_js_bridge : string
 val file_listen_transactions : string
