+++ src/tezos_interop/tezos_js_bridge.js
@@ -1,12 +1,15 @@
 "use strict";
 
 const { TezosToolkit } = require("@taquito/taquito");
+const { RpcClient } = require("@taquito/rpc");
 const { InMemorySigner } = require("@taquito/signer");
 const { pipeline, Transform } = require("stream");
+const { inspect } = require("util");
 
 /**
  * @typedef TransactionRequest
  * @type {object}
+ * @property {"transaction"} kind
  * @property {string} rpc_node
  * @property {string} secret
  * @property {number} confirmation
@@ -14,10 +17,17 @@ const { pipeline, Transform } = require("stream");
  * @property {string} entrypoint
  * @property {object} payload
 
+ * @typedef StorageRequest
+ * @type {object}
+ * @property {"storage"} kind
+ * @property {string} rpc_node
+ * @property {number} confirmation
+ * @property {string} destination
+
  * @typedef Request
  * @type {object}
  * @property {number} id
- * @property {TransactionRequest} content
+ * @property {TransactionRequest | StorageRequest} content
 
  * @typedef ErrorResponse
  * @type {object}
@@ -29,10 +39,15 @@ const { pipeline, Transform } = require("stream");
  * @property {"applied" | "failed" | "skipped" | "backtracked" | "unknown"} status
  * @property {string} hash
 
+ * @typedef StorageResponse
+ * @type {object}
+ * @property {"success"} status
+ * @property {rpc.MichelsonV1Expression} storage
+
  * @typedef Response
  * @type {object}
  * @property {number} id
- * @property {TransactionResponse | ErrorResponse} content
+ * @property {TransactionResponse | StorageResponse | ErrorResponse} content
  */
 
 const failure = (err) => {
@@ -101,7 +116,7 @@ const config = {
 };
 
 /** @param {TransactionRequest} request */
-const onRequest = async (request) => {
+const onTransactionRequest = async (request) => {
   const { rpc_node, secret, confirmation, destination, entrypoint, payload } =
     request;
 
@@ -121,13 +136,56 @@ const onRequest = async (request) => {
   return { status, hash };
 };
 
+/** @param {StorageRequest} request */
+const onStorageRequest = async (request) => {
+  const { rpc_node, confirmation, destination } = request;
+  const client = new RpcClient(rpc_node);
+  const Tezos = new TezosToolkit(rpc_node);
+  Tezos.setProvider({ config });
+
+  const block = await client.getBlock(); // fetches the head
+  const contract = await client.getContract(destination, {
+    block: block.hash,
+  });
+  const storage = contract.script.storage;
+  /* To make sure the storage state is finalized, we query the last
+     block and any one of its operation, and wait till it receives
+     `n` confirmations (where n is the minimum blocks needs to
+     consider reorg highly unlikely. ie finality) */
+  const operationFromHead = block.operations.flat()[0];
+  if (!operationFromHead) {
+    throw new Error("Internal error: operationFromHead was undefined");
+  }
+
+  const operation = await Tezos.operation.createTransactionOperation(
+    operationFromHead.hash
+  );
+  const result = await operation.confirmation(confirmation);
+  if (!(await result.isInCurrentBranch())) {
+    throw new Error("Not in current Branch");
+  }
+
+  return { status: "success", storage };
+};
+
+/** @param {Request} request */
+const onRequest = async (request) => {
+  if (request.kind === "transaction") {
+    return onTransactionRequest(request);
+  } else if (request.kind === "storage") {
+    return onStorageRequest(request);
+  } else {
+    failure(new Error("invalid request.kind: " + JSON.stringify(request.kind)));
+  }
+};
+
 read((request) => {
   const { id, content } = request;
   onRequest(content)
     .then((content) => write({ id, content }))
     .catch((err) => {
       const status = "error";
-      const error = JSON.stringify(err);
+      const error = inspect(err);
       const content = { status, error };
       write({ id, content });
     })
