+++ src/tezos_interop/webpack.config.js
@@ -13,7 +13,6 @@ module.exports = (env) => ({
   ...resolve,
   entry: {
     tezos_js_bridge: "./tezos_js_bridge.js",
-    fetch_storage: "./fetch_storage.js",
     listen_transactions: "./listen_transactions.js",
   },
   output: {
